jewelry_classifier/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
torch==2.6.0
torchvision==0.21.0
timm==1.0.9
Pillow==10.4.0
pydantic==2.9.2
pydantic-settings==2.5.2